fix: Round service finish delay to next or equal minute in vehicle routing
When service finish delay is positive and less than a minute, a hard
constraint match with 0 weight would be generated (since service finish
delay was rounded down to zero).

This make service finish delay round up, so a delay of one second
would be rounded up to 1 minute, and have a penalty of 1 instead
of zero.